requirements.txt
python-telegram-bot==21.10
tvdatafeed @git+https://github.com/rongardF/tvdatafeed.git
mplfinance==0.12.10b0
pandas==2.2.3
tvdatafeed==2.1.0
telegram==0.0.1
python-dotenv==1.0.1
googlefinance==0.7